+++ kubejs/client_scripts/ponder/modern_industrialization/large_steam_macerator.js
@@ -1,0 +1,45 @@
+// -----------------------------------------
+// CREATED BY STATIC FOR USE IN
+// STATECH INDUSTRY
+// -----------------------------------------
+
+Ponder.registry((event) => {
+    event.create('modern_industrialization:large_steam_macerator').scene('large_steam_macerator', 'The Large Steam Macerator', 'kubejs:large_steam_macerator', (scene, util) => {
+        // Bottom layer
+        for (let x = 0; x < 5; x++) {
+            for (let z = 0; z < 5; z++) {
+                scene.world.showSection([x, 0, z], Facing.DOWN);
+            }
+        }
+
+        // First two layers
+        for (let y = 1; y < 3; y++) {
+            for (let x = 0; x < 5; x++) {
+                for (let z = 0; z < 5; z++) {
+                    scene.world.showSection([x, y, z], Facing.DOWN);
+                }
+                scene.idle(3);
+            }
+            scene.idle(5);
+        }
+
+        scene
+            .text('60', 'The center is made of Bronze Machine Pipe Casings', [2.0, 3.0, 2.0])
+            .colored(PonderPalette.WHITE)
+            .placeNearTarget()
+            .attachKeyFrame();
+        scene.idle(80);
+
+        // Build the rest
+        for (let y = 3; y < 5; y++) {
+            for (let x = 0; x < 5; x++) {
+                for (let z = 0; z < 5; z++) {
+                    scene.world.showSection([x, y, z], Facing.DOWN);
+                }
+                scene.idle(3);
+            }
+            scene.idle(5);
+        }
+        scene.idle(20);
+    });
+});